refactor: 피드 관련 예외 클래스 위치 변경
