+++ src/main/java/com/sprint/ootd5team/domain/feed/service/FeedServiceImpl.java
@@ -7,8 +7,8 @@ import com.sprint.ootd5team.domain.feed.dto.request.FeedListRequest;
 import com.sprint.ootd5team.domain.feed.dto.request.FeedUpdateRequest;
 import com.sprint.ootd5team.domain.feed.dto.response.FeedDtoCursorResponse;
 import com.sprint.ootd5team.domain.feed.entity.Feed;
-import com.sprint.ootd5team.domain.feed.exception.FeedNotFoundException;
-import com.sprint.ootd5team.domain.feed.exception.InvalidSortOptionException;
+import com.sprint.ootd5team.base.exception.feed.FeedNotFoundException;
+import com.sprint.ootd5team.base.exception.feed.InvalidSortOptionException;
 import com.sprint.ootd5team.domain.feed.repository.feed.FeedRepository;
 import com.sprint.ootd5team.domain.feed.repository.feedClothes.FeedClothesRepository;
 import java.util.List;
